% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calorimetry.R
\name{aggregate_calorimetry}
\alias{aggregate_calorimetry}
\title{Aggregate Calorimetry Data}
\usage{
aggregate_calorimetry(calo, by = NULL, fun = mean)
}
\arguments{
\item{calo}{A CalorimetryData object}

\item{by}{Character vector of grouping columns (uses id_col, time_col, protocol_col by default)}

\item{fun}{Aggregation function (default: mean)}
}
\value{
A CalorimetryData object with aggregated data
}
\description{
Aggregate calorimetry measurements by subject, time, and optionally protocol.
}
