% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxidation.R
\name{analyze_oxidation}
\alias{analyze_oxidation}
\title{Analyze Oxidation (Main Workflow)}
\usage{
analyze_oxidation(
  study,
  time_range = NULL,
  aggregate = FALSE,
  calc_energy = TRUE,
  control_protocol = NULL
)
}
\arguments{
\item{study}{An OxidationStudy object}

\item{time_range}{Optional time range to filter (vector of 2)}

\item{aggregate}{Logical, aggregate by time point (default: FALSE)}

\item{calc_energy}{Logical, calculate energy contributions (default: TRUE)}

\item{control_protocol}{Name of control protocol for Rref (required if isotopes present)}
}
\value{
An OxidationResults S7 object
}
\description{
Main analysis function that performs complete oxidation analysis.
}
\examples{
\dontrun{
study <- oxidation_study(
  calorimetry = calo_data,
  isotopes = iso_data,
  subjects = subject_data
)
results <- analyze_oxidation(study, time_range = c(30, 120))
}
}
