% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_numeric_time}
\alias{as_numeric_time}
\title{Convert Time to Numeric}
\usage{
as_numeric_time(x)
}
\arguments{
\item{x}{Time vector (can be factor, character, or numeric)}
}
\value{
Numeric vector
}
\description{
Safely convert a time column to numeric.
}
