% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy.R
\name{calc_energy_percent}
\alias{calc_energy_percent}
\title{Calculate Energy Percentages}
\usage{
calc_energy_percent(energy)
}
\arguments{
\item{energy}{A data frame with energy yield columns (from calc_energy_yield)}
}
\value{
A tibble with percentage contribution columns added
}
\description{
Calculate the percentage contribution of each substrate to total energy.
}
\examples{
\dontrun{
ox <- calc_substrate_oxidation(calo)
energy <- calc_energy_yield(ox)
energy_pct <- calc_energy_percent(energy)
}
}
