% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopes.R
\name{calc_fexo}
\alias{calc_fexo}
\title{Calculate Fraction Exogenous (Fexo)}
\usage{
calc_fexo(isotopes, rref)
}
\arguments{
\item{isotopes}{An IsotopeData object (must include rpla)}

\item{rref}{Reference enrichment value or data frame}
}
\value{
A tibble with Fexo values (\%)
}
\description{
Calculate the fraction of plasma glucose derived from exogenous sources.
}
\details{
Fexo = ((Rpla - Rref) / (Rexo - Rref)) * 100

Where Rpla is the 13C enrichment of plasma glucose.
}
