% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein.R
\name{calc_protein_gas_exchange}
\alias{calc_protein_gas_exchange}
\title{Calculate Protein Contribution to Gas Exchange}
\usage{
calc_protein_gas_exchange(
  protein_ox,
  vo2_factor = 1.0075,
  vco2_factor = 0.8443
)
}
\arguments{
\item{protein_ox}{Protein oxidation rate (g/min) - vector or data frame}

\item{vo2_factor}{VO2 per gram protein (default: 1.0075 L/g, derived from Institute of Medicine, 2005, as presented in Telmosse, 2022)}

\item{vco2_factor}{VCO2 per gram protein (default: 0.8443 L/g, derived from Institute of Medicine, 2005, as presented in Telmosse, 2022)}
}
\value{
A tibble with VO2p and VCO2p values
}
\description{
Calculate the VO2 and VCO2 attributable to protein oxidation.
}
