% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calorimetry.R
\name{calc_rer}
\alias{calc_rer}
\title{Calculate RER from Calorimetry Data}
\usage{
calc_rer(calo, vo2_col = NULL, vco2_col = NULL)
}
\arguments{
\item{calo}{A CalorimetryData object or data frame}

\item{vo2_col}{Name of the VO2 column (if data frame)}

\item{vco2_col}{Name of the VCO2 column (if data frame)}
}
\value{
A numeric vector of RER values
}
\description{
Calculate Respiratory Exchange Ratio (VCO2/VO2).
}
