% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein.R
\name{calc_sweat_loss}
\alias{calc_sweat_loss}
\title{Calculate Sweat Loss}
\usage{
calc_sweat_loss(subjects, calo, environment, exercise_duration = 120)
}
\arguments{
\item{subjects}{A SubjectData object or data frame with mass/fluid data}

\item{calo}{A CalorimetryData object for respiratory calculations}

\item{environment}{An EnvironmentData object for vapor pressure calculations}

\item{exercise_duration}{Exercise duration in minutes (default: 120)}
}
\value{
A tibble with sweat loss estimates (kg)
}
\description{
Estimate sweat loss from mass balance considering fluid intake and
respiratory gas/water losses.
}
