% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy.R
\name{calc_total_energy}
\alias{calc_total_energy}
\title{Calculate Total Energy Expenditure}
\usage{
calc_total_energy(vo2, vco2, method = c("weir", "brockway"))
}
\arguments{
\item{vo2}{Oxygen consumption (L/min)}

\item{vco2}{Carbon dioxide production (L/min)}

\item{method}{Calculation method: "weir" or "brockway" (default: "weir")}
}
\value{
Energy expenditure (kcal/min)
}
\description{
Calculate total energy expenditure from VO2 using the Weir equation.
}
\details{
Weir equation (1949):
EE (kcal/min) = 3.941 * VO2 + 1.106 * VCO2

Brockway equation (1987):
EE (kcal/min) = 3.869 * VO2 + 1.195 * VCO2
}
