% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein.R
\name{calc_urea_loss}
\alias{calc_urea_loss}
\title{Calculate Urea Loss}
\usage{
calc_urea_loss(urea, sweat_loss, subjects = NULL)
}
\arguments{
\item{urea}{A UreaData object}

\item{sweat_loss}{A data frame with sweat loss (from calc_sweat_loss)}

\item{subjects}{A SubjectData object (for urine volume)}
}
\value{
A tibble with urea loss components
}
\description{
Calculate total urea nitrogen loss from sweat and urine.
}
