% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{CalorimetryData}
\alias{CalorimetryData}
\title{Calorimetry Data Class}
\arguments{
\item{data}{A data frame containing calorimetry measurements}

\item{id_col}{Name of the subject ID column (default: "id")}

\item{time_col}{Name of the time column (default: "time")}

\item{vo2_col}{Name of the VO2 column (default: "vo2")}

\item{vco2_col}{Name of the VCO2 column (default: "vco2")}

\item{vo2_unit}{Unit for VO2/VCO2: "L/min" or "mL/min" (default: "L/min")}

\item{protocol_col}{Name of the protocol/condition column (default: NULL)}
}
\value{
A CalorimetryData S7 object
}
\description{
S7 class for storing indirect calorimetry measurements (VO2, VCO2, RER).
}
\examples{
df <- data.frame(
  id = c(1, 1, 2, 2),
  time = c(30, 60, 30, 60),
  vo2 = c(2.5, 2.6, 2.4, 2.5),
  vco2 = c(2.3, 2.4, 2.2, 2.3)
)
calo <- CalorimetryData(data = df)
}
