% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{create_report_template}
\alias{create_report_template}
\title{Create Report Template}
\usage{
create_report_template(dest = "oxidation_report.qmd", overwrite = FALSE)
}
\arguments{
\item{dest}{Destination file path}

\item{overwrite}{Overwrite existing file}
}
\value{
Invisible path to created file
}
\description{
Copy the default report template to a location for customization.
}
