% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxidizr-package.R
\docType{data}
\name{energy_factors}
\alias{energy_factors}
\title{Energy conversion factors}
\format{
A named list with three elements:
\describe{
\item{cho}{Carbohydrate energy factor (default: 3.72 kcal/g)}
\item{fat}{Fat energy factor (default: 9.44 kcal/g)}
\item{protein}{Protein energy factor (default: 4.70 kcal/g)}
}
}
\usage{
energy_factors
}
\description{
Default energy conversion factors (kcal/g) for substrate oxidation
}
\keyword{datasets}
