% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\docType{data}
\name{environment_thresholds}
\alias{environment_thresholds}
\title{Environment Validation Thresholds}
\format{
A named list with thresholds for temperature, humidity, and pressure
}
\usage{
environment_thresholds
}
\description{
Default thresholds for validating environmental conditions.
}
\references{
ACSM. (2022). ACSM's Guidelines for Exercise Testing and Prescription.
}
\keyword{datasets}
