% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{EnvironmentData}
\alias{EnvironmentData}
\title{Environment Data Class}
\arguments{
\item{data}{Data frame with environmental measurements}

\item{id_col}{Name of the subject ID column}

\item{temp_col}{Name of the temperature column (Celsius)}

\item{humidity_col}{Name of the relative humidity column (\%)}

\item{pressure_col}{Name of the atmospheric pressure column (mmHg)}
}
\value{
An EnvironmentData S7 object
}
\description{
S7 class for storing environmental conditions during testing.
}
