% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calorimetry.R
\name{filter_time_range}
\alias{filter_time_range}
\title{Filter Calorimetry Data by Time}
\usage{
filter_time_range(calo, time_range)
}
\arguments{
\item{calo}{A CalorimetryData object}

\item{time_range}{Numeric vector of length 2 specifying min and max time}
}
\value{
A CalorimetryData object with filtered data
}
\description{
Filter calorimetry data to a specific time range.
}
