% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_mean_sd}
\alias{format_mean_sd}
\title{Format Mean +/- SD}
\usage{
format_mean_sd(mean, sd, decimals = 2, sep = " ± ")
}
\arguments{
\item{mean}{Mean value}

\item{sd}{Standard deviation}

\item{decimals}{Number of decimal places}

\item{sep}{Separator (default: " +/- ")}
}
\value{
Character string
}
\description{
Format a mean and standard deviation as "mean +/- sd".
}
