% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_data}
\alias{get_data}
\title{Extract data from oxidizr objects}
\usage{
get_data(x, ...)
}
\arguments{
\item{x}{An oxidizr S7 object}

\item{...}{Additional arguments}
}
\value{
A tibble/data frame
}
\description{
Generic function to extract the underlying data frame from oxidizr S7 objects.
}
