% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\docType{data}
\name{isotope_thresholds}
\alias{isotope_thresholds}
\title{Isotope Validation Thresholds}
\format{
A named list with thresholds for each isotope measurement
}
\usage{
isotope_thresholds
}
\description{
Default thresholds for validating 13C isotope enrichment data.
}
\references{
Craig H. (1957). Isotopic standards for carbon and oxygen and correction
factors for mass-spectrometric analysis. Geochimica et Cosmochimica Acta.
Pirnay F, et al. (1977). Fate of exogenous glucose during exercise of
different intensities in humans. J Appl Physiol.
}
\keyword{datasets}
