% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_columns}
\alias{list_columns}
\title{List Available Columns}
\usage{
list_columns(results)
}
\arguments{
\item{results}{An OxidationResults object}
}
\value{
Invisible NULL (prints to console)
}
\description{
Print available columns in oxidation results for reference.
}
