% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{OxidationResults}
\alias{OxidationResults}
\title{Oxidation Results Class}
\arguments{
\item{oxidation_rates}{Data frame with oxidation rates (g/min)}

\item{energy_contributions}{Data frame with energy contributions (kcal and \%)}

\item{partition}{Data frame with CHO partitioning (exo/endo/mus/liv)}

\item{study}{The original OxidationStudy object}

\item{settings}{List of analysis settings used}
}
\value{
An OxidationResults S7 object
}
\description{
S7 class for storing the results of oxidation analysis.
}
