% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxidizr-package.R
\docType{package}
\name{oxidizr-package}
\alias{oxidizr}
\alias{oxidizr-package}
\title{oxidizr: Substrate Oxidation Analysis During Exercise}
\description{
Analyze substrate oxidation during exercise using indirect calorimetry and stable isotope tracers. Calculates carbohydrate, fat, and protein oxidation rates with optional partitioning of exogenous vs endogenous carbohydrate sources using 13C isotope methods. Provides S7-based classes for data management, publication-ready visualizations, and formatted tables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jtremblay/oxidizr}
  \item Report bugs at \url{https://github.com/jtremblay/oxidizr/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Tremblay \email{jonathan.tremblay@umontreal.ca}

}
\keyword{internal}
