% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_cho_partition}
\alias{plot_cho_partition}
\title{Plot CHO Partition}
\usage{
plot_cho_partition(results, by = "protocol", type = c("exo_endo", "mus_liv"))
}
\arguments{
\item{results}{An OxidationResults object}

\item{by}{Grouping variable}

\item{type}{Type of partition: "exo_endo" or "mus_liv"}
}
\value{
A ggplot2 object
}
\description{
Create a plot showing the partitioning of CHO oxidation sources.
}
