% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_energy_contribution}
\alias{plot_energy_contribution}
\title{Plot Energy Contribution}
\usage{
plot_energy_contribution(
  results,
  by = "protocol",
  partition = FALSE,
  show_error = TRUE
)
}
\arguments{
\item{results}{An OxidationResults object}

\item{by}{Grouping variable (typically "protocol")}

\item{partition}{Show CHO partition (exo/endo) instead of total}

\item{show_error}{Show error bars}
}
\value{
A ggplot2 object
}
\description{
Create a stacked bar chart of energy contributions by substrate.
}
