% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_isotope_enrichment}
\alias{plot_isotope_enrichment}
\title{Plot Isotope Enrichment}
\usage{
plot_isotope_enrichment(
  results,
  type = c("expired", "plasma"),
  by = NULL,
  time_col = "time",
  show_error = TRUE
)
}
\arguments{
\item{results}{An OxidationResults object with isotope data}

\item{type}{Type of enrichment: "expired" (Rexp), "plasma" (Rpla)}

\item{by}{Grouping variable}

\item{time_col}{Name of time column}

\item{show_error}{Show error bars}
}
\value{
A ggplot2 object
}
\description{
Create a time course plot of 13C isotope enrichment.
}
