% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_vo2_timecourse}
\alias{plot_vo2_timecourse}
\title{Plot VO2 Time Course}
\usage{
plot_vo2_timecourse(
  results,
  by = NULL,
  time_col = "time",
  show_error = TRUE,
  error_type = c("sd", "se")
)
}
\arguments{
\item{results}{An OxidationResults object or data frame}

\item{by}{Grouping variable for color}

\item{time_col}{Name of time column}

\item{show_error}{Show error bars}

\item{error_type}{Type of error bars}
}
\value{
A ggplot2 object
}
\description{
Create a time course plot of oxygen consumption.
}
