% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quick_summary}
\alias{quick_summary}
\title{Quick Summary of Oxidation Results}
\usage{
quick_summary(results, by = NULL)
}
\arguments{
\item{results}{An OxidationResults object or data frame}

\item{by}{Optional grouping variable}
}
\value{
Invisible NULL (prints summary to console)
}
\description{
Print a quick summary of key oxidation metrics.
}
