% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxidizr-package.R
\docType{data}
\name{stoich_coefficients}
\alias{stoich_coefficients}
\title{Stoichiometric coefficients}
\format{
A named list with coefficients for CHO and fat calculations
}
\usage{
stoich_coefficients
}
\description{
Coefficients for calculating substrate oxidation from gas exchange,
as derived and presented in Telmosse (2022).
}
\references{
Telmosse, E. (2022). Contribution des glucides exogènes à la fourniture d’énergie lors d’un effort prolongé : analyse par régression multiple. Mémoire de maîtrise, Université de Montréal.
Livesey, G., & Elia, M. (1988). Estimation of energy expenditure, net carbohydrate utilization, and net fat oxidation and synthesis by indirect calorimetry: evaluation of errors with special reference to the detailed composition of fuels. \emph{The American Journal of Clinical Nutrition}, \emph{47}(4), 608–628.
Institute of Medicine, Food and Nutrition Board. (2005). \emph{Dietary Reference Intakes for Energy, Carbohydrate, Fiber, Fat, Fatty Acids, Cholesterol, Protein, and Amino Acids}. National Academies Press.
Péronnet, F., & Massicotte, D. (1991). Table of nonprotein respiratory quotient: an update. \emph{Canadian Journal of Sport Sciences = Journal Canadien Des Sciences Du Sport}, \emph{16}(1), 23–29.
}
\keyword{datasets}
