% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{SubjectData}
\alias{SubjectData}
\title{Subject Data Class}
\arguments{
\item{data}{Data frame with subject information}

\item{id_col}{Name of the subject ID column}

\item{protocol_col}{Name of the protocol/condition column (optional)}

\item{mass_cols}{Named list of mass columns (initial, final)}

\item{fluid_cols}{Named list of fluid columns (water, saline, urine)}
}
\value{
A SubjectData S7 object
}
\description{
S7 class for storing subject characteristics and daily measurements.
}
