% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tbl_energy_contribution}
\alias{tbl_energy_contribution}
\title{Energy Contribution Table}
\usage{
tbl_energy_contribution(
  results,
  by = NULL,
  show_kcal = TRUE,
  show_percent = TRUE,
  decimals = 1
)
}
\arguments{
\item{results}{An OxidationResults object}

\item{by}{Grouping variable}

\item{show_kcal}{Show absolute energy (kcal/min)}

\item{show_percent}{Show percentage contributions}

\item{decimals}{Number of decimal places}
}
\value{
A gt table object
}
\description{
Create a table of energy contributions by substrate.
}
