% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tbl_oxidation_summary}
\alias{tbl_oxidation_summary}
\title{Oxidation Summary Table}
\usage{
tbl_oxidation_summary(results, by = NULL, vars = NULL, decimals = 2)
}
\arguments{
\item{results}{An OxidationResults object}

\item{by}{Grouping variable (e.g., "protocol")}

\item{vars}{Variables to include (NULL = auto)}

\item{decimals}{Number of decimal places}
}
\value{
A gt table object
}
\description{
Create a summary table of oxidation rates.
}
