% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tbl_stats}
\alias{tbl_stats}
\title{Statistical Results Table}
\usage{
tbl_stats(model, type = c("anova", "posthoc"), decimals = 3)
}
\arguments{
\item{model}{A fitted model object (aov, lmer, etc.)}

\item{type}{Type of table: "anova" or "posthoc"}

\item{decimals}{Number of decimal places}
}
\value{
A gt table object
}
\description{
Create a table of statistical test results.
}
