% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tbl_subjects}
\alias{tbl_subjects}
\title{Subject Characteristics Table}
\usage{
tbl_subjects(study, by = NULL, vars = NULL, decimals = 1)
}
\arguments{
\item{study}{An OxidationStudy object}

\item{by}{Grouping variable (e.g., "protocol")}

\item{vars}{Variables to include (NULL = auto-detect)}

\item{decimals}{Number of decimal places}
}
\value{
A gt table object
}
\description{
Create a formatted table of subject characteristics.
}
