% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\docType{data}
\name{urea_thresholds}
\alias{urea_thresholds}
\title{Urea Validation Thresholds}
\format{
A named list with thresholds for sweat and urine urea
}
\usage{
urea_thresholds
}
\description{
Default thresholds for validating urea concentration data.
}
\references{
Shirreffs SM, Maughan RJ. (1997). Whole body sweat collection in humans.
J Appl Physiol, 82(1), 336-341.
Lemon PW. (1991). Effect of exercise on protein requirements. J Sports Sci.
}
\keyword{datasets}
