% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{UreaData}
\alias{UreaData}
\title{Urea Data Class}
\arguments{
\item{data}{Data frame with urea concentrations}

\item{id_col}{Name of the subject ID column}

\item{sweat_urea_col}{Name of the sweat urea concentration column (mmol/L)}

\item{urine_urea_col}{Name of the urine urea concentration column (mmol/L)}

\item{urine_vol_col}{Name of the urine volume column (mL)}
}
\value{
A UreaData S7 object
}
\description{
S7 class for storing urea concentration data for protein oxidation estimation.
}
